o Use 'pvcreate --setphysicalvolumesize' with no short form (instead of -s)   and add severe warning if it's used to make a device seem bigger than   it really is.  This is not an option people should be using as it   breaks metadata integrity. o Use uint64_t throughout (rather than unsigned long long) o Convert a few messages that contain pathnames into the more common form:   pathname: message

+++ lib/display/display.h
@@ -28,7 +28,7 @@
 typedef	enum {SIZE_LONG=0, SIZE_SHORT=1} size_len_t;
 
 /* Specify size in KB */
-char *display_size(unsigned long long size, size_len_t sl);
+char *display_size(uint64_t size, size_len_t sl);
 char *display_uuid(char *uuidstr);
 
 void pvdisplay_colons(struct physical_volume *pv);
